requirements.txt
openai>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.1
pytest>=8.2.0
streamlit>=1.36.0
google-generativeai>=0.7.0

